pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "steep-digest"
version = "0.1.0"
description = "Personal newsletter digest: Gmail allowlist → Claude → Gmail + Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-api-python-client>=2.136.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "anthropic>=0.42.0",
  "httpx>=0.27.0",
  "html2text>=2024.2.26",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[project.scripts]
steep-digest = "steep_digest.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
